Implement storage of memories via frontend->backend.

+++ backend/src/models/memory/memory.go
@@ -2,7 +2,7 @@ package memory
 
 type Memory struct {
 	Text      string
-	Timestamp int32
+	Timestamp int64
 	Latitude  float32
 	Longitude float32
 }
